+++ core/src/main/java/org/seedstack/seed/core/internal/logging/LoggingTypeListener.java
@@ -8,6 +8,8 @@
 
 package org.seedstack.seed.core.internal.logging;
 
+import static java.lang.String.format;
+
 import com.google.inject.TypeLiteral;
 import com.google.inject.spi.TypeEncounter;
 import com.google.inject.spi.TypeListener;
@@ -15,6 +17,8 @@ import java.lang.reflect.Field;
 import java.util.HashSet;
 import java.util.Set;
 import org.seedstack.seed.Logging;
+import org.seedstack.seed.SeedException;
+import org.seedstack.seed.core.internal.CoreErrorCode;
 import org.slf4j.Logger;
 
 /**
@@ -26,8 +30,15 @@ class LoggingTypeListener implements TypeListener {
         Set<Field> fields = new HashSet<>();
         for (Class<?> c = typeLiteral.getRawType(); c != Object.class; c = c.getSuperclass()) {
             for (Field field : c.getDeclaredFields()) {
-                if (field.getType() == Logger.class && field.isAnnotationPresent(Logging.class)) {
-                    fields.add(field);
+                if (field.isAnnotationPresent(Logging.class)) {
+                    if (field.getType() == Logger.class) {
+                        fields.add(field);
+                    } else {
+                        throw SeedException.createNew(CoreErrorCode.BAD_LOGGER_TYPE)
+                                .put("field", format("%s.%s", field.getDeclaringClass().getName(), field.getName()))
+                                .put("expectedType", Logger.class.getName())
+                                .put("givenType", field.getType().getName());
+                    }
                 }
             }
         }